native-port: OSRoundUp/Down32B macros + JKRHeap arena size use uintptr_t
OSRoundUp32B/OSRoundDown32B cast their arg (often a pointer) to u32 — truncates under
LP64. Use uintptr_t so pointer alignment is correct on x86-64/arm64. JKRHeap::initArena
computes arena size from two pointers losslessly via uintptr_t. (JKRHeap/JUTException
also contain inline asm → full clean-up lands with the asm-reimplementation pass.)

+++ src/JSystem/JKernel/JKRHeap.cpp
@@ -69,7 +69,7 @@ bool JKRHeap::initArena(char** outUserRamStart, u32* outUserRamSize,
 	OSSetArenaLo(arenaHi);
 	OSSetArenaHi(arenaHi);
 	*outUserRamStart = (char*)arenaLo;
-	*outUserRamSize  = (u32)arenaHi - (u32)arenaLo;
+	*outUserRamSize  = (u32)((uintptr_t)arenaHi - (uintptr_t)arenaLo);
 	OSReport("arenaLo=%x arenaHi=%x \n", arenaLo, arenaHi);
 	return true;
 }
